+++ packages/fastforge/bin/main.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
 
-import 'package:unified_distributor/unified_distributor.dart';
+import 'package:fastforge/fastforge.dart';
 
 Future<void> main(List<String> args) async {
   try {