+++ packages/fastforge/test/src/extensions/string_test.dart
@@ -1,5 +1,5 @@
+import 'package:fastforge/fastforge.dart';
 import 'package:test/test.dart';
-import 'package:unified_distributor/src/extensions/string.dart';
 
 void main() {
   // print('black        '.black());