+++ packages/flutter_distributor/lib/flutter_distributor.dart
@@ -1,6 +1,6 @@
 library flutter_distributor;
 
-import 'package:unified_distributor/unified_distributor.dart';
+import 'package:fastforge/fastforge.dart';
 
 class FlutterDistributor extends UnifiedDistributor {
   FlutterDistributor() : super('flutter_distributor', 'Flutter Distributor');